+++ src/common/win32/wstring.h
@@ -1,16 +1,18 @@
 /*
- * ceph-dokan - Win32 CephFS client based on Dokan
+ * Ceph - scalable distributed file system
  *
- * Copyright (C) 2021 SUSE LINUX GmbH
+ * Copyright (C) 2022 Cloudbase Solutions
  *
  * This is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License version 2.1, as published by the Free Software
  * Foundation.  See file COPYING.
  *
-*/
+ */
 
-#include "include/compat.h"
+#pragma once
 
-void to_filetime(time_t t, LPFILETIME pft);
-void to_unix_time(FILETIME ft, time_t *t);
+#include <string>
+
+std::wstring to_wstring(const std::string& str);
+std::string to_string(const std::wstring& wstr);
