+++ src/dokan/utils.cc
@@ -12,20 +12,6 @@
 
 #include "utils.h"
 
-#include <boost/locale/encoding_utf.hpp>
-
-using boost::locale::conv::utf_to_utf;
-
-std::wstring to_wstring(const std::string& str)
-{
-  return utf_to_utf<wchar_t>(str.c_str(), str.c_str() + str.size());
-}
-
-std::string to_string(const std::wstring& str)
-{
-  return utf_to_utf<char>(str.c_str(), str.c_str() + str.size());
-}
-
 void to_filetime(time_t t, LPFILETIME pft)
 {
   // Note that LONGLONG is a 64-bit value
