+++ src/tools/rbd_wnbd/rbd_wnbd.cc
@@ -20,8 +20,6 @@
 #include <sys/socket.h>
 #include <unistd.h>
 
-#include <boost/locale/encoding_utf.hpp>
-
 #include "wnbd_handler.h"
 #include "rbd_wnbd.h"
 
@@ -38,6 +36,7 @@
 #include "common/errno.h"
 #include "common/version.h"
 #include "common/win32/service.h"
+#include "common/win32/wstring.h"
 #include "common/admin_socket_client.h"
 
 #include "global/global_init.h"
@@ -54,17 +53,6 @@
 #define dout_prefix *_dout << "rbd-wnbd: "
 
 using namespace std;
-using boost::locale::conv::utf_to_utf;
-
-std::wstring to_wstring(const std::string& str)
-{
-  return utf_to_utf<wchar_t>(str.c_str(), str.c_str() + str.size());
-}
-
-std::string to_string(const std::wstring& str)
-{
-  return utf_to_utf<char>(str.c_str(), str.c_str() + str.size());
-}
 
 bool is_process_running(DWORD pid)
 {
